+++ src/main/java/org/test/task/service/UsersStorageService.java
@@ -1,16 +1,14 @@
 package org.test.task.service;
 
-import org.springframework.stereotype.Service;
-import org.test.task.model.Message;
 import org.test.task.model.User;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class UsersStorageService {
 
-    private static final AtomicLong counter = new AtomicLong();
+    private static final AtomicInteger counter = new AtomicInteger();
 
     public static List<User> users = new ArrayList<User>();
